requirements.txt
streamlit
tensorflow
pillow
numpy
googletrans==4.0.0-rc1